Let a run that waited for the build lock reuse the winner's work
Fifth round of Codex review.

- Two runs that both found the assets stale each took their turn and then
  built the same thing, the second one for nothing. A run that has waited
  now asks again, once it is its turn, and usually finds the run ahead of
  it has just made these very assets. The common case is untouched: if
  the assets already look current, the answer comes from a handful of
  stat calls and the lock is never taken. Measured on two runs started
  together with a cold cache: the waiter finishes ~200ms after the
  winner, not a whole build later.
- Modified times are compared at second granularity now. Some
  filesystems only keep them to the nearest second, which could round a
  save made just after the build began to just before it, and let a
  raced build be recorded. Erring a second early only ever costs a
  rebuild.
- The mutation runner's note still said the harness cleans up generated
  assets on an interrupted run. It deliberately keeps them now.
- Split the cache tests: the ones about the build session moved in with
  the other session tests, putting both files back under 400 lines.

Taking the lock twice in one process would wait on itself forever, so
the shared writer now exposes its unlocked body for the one caller that
holds the lock already. Both entry points take it exactly once.
