+++ scripts/static-assets/session.ts
@@ -62,12 +62,17 @@ export const deferStaticAssetBuild = (
  * The build a run should work with: when the assets on disk are already
  * current, one that waits until something actually asks; otherwise the real
  * build, run now, because the assets have to exist before any test starts.
+ *
+ * `buildLater` covers the case where a build put off until something asks for
+ * it needs to be started differently from one run right now — the caller may be
+ * holding a lock now that it will not hold then. It defaults to `build`.
  */
 export const buildOrReuseStaticAssets = async (
   upToDate: boolean,
   build: () => Promise<StaticAssetBuild>,
+  buildLater: () => Promise<StaticAssetBuild> = build,
 ): Promise<StaticAssetBuild> =>
-  upToDate ? deferStaticAssetBuild(build) : await build();
+  upToDate ? deferStaticAssetBuild(buildLater) : await build();
 
 export const fileExists = async (file: string): Promise<boolean> => {
   try {
